[Chore] #66 - review와 관련된 패키지 이동
