+++ src/main/java/com/arabook/arabook/review/controller/ReviewController.java
@@ -1,4 +1,4 @@
-package com.arabook.arabook.api.review.controller;
+package com.arabook.arabook.review.controller;
 
 import static com.arabook.arabook.common.success.review.ReviewSuccessType.*;
 
@@ -17,13 +17,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.arabook.arabook.api.global.security.AuthMember;
-import com.arabook.arabook.api.review.controller.dto.request.CreateReviewRequest;
-import com.arabook.arabook.api.review.controller.dto.request.UpdateReviewRequest;
-import com.arabook.arabook.api.review.controller.dto.response.ReviewDetailResponse;
-import com.arabook.arabook.api.review.controller.dto.response.ReviewIdResponse;
-import com.arabook.arabook.api.review.controller.dto.response.ReviewsResponse;
-import com.arabook.arabook.api.review.service.ReviewService;
 import com.arabook.arabook.common.response.ResponseTemplate;
+import com.arabook.arabook.review.controller.dto.request.CreateReviewRequest;
+import com.arabook.arabook.review.controller.dto.request.UpdateReviewRequest;
+import com.arabook.arabook.review.controller.dto.response.ReviewDetailResponse;
+import com.arabook.arabook.review.controller.dto.response.ReviewIdResponse;
+import com.arabook.arabook.review.controller.dto.response.ReviewsResponse;
+import com.arabook.arabook.review.service.ReviewService;
 
 import lombok.RequiredArgsConstructor;
