+++ src/main/java/com/arabook/arabook/review/entity/Review.java
@@ -1,4 +1,4 @@
-package com.arabook.arabook.storage.domain.review.entity;
+package com.arabook.arabook.review.entity;
 
 import static com.arabook.arabook.common.exception.review.ReviewExceptionType.*;
 
@@ -19,8 +19,8 @@ import jakarta.validation.constraints.NotNull;
 import com.arabook.arabook.book.entity.Book;
 import com.arabook.arabook.common.exception.review.ReviewException;
 import com.arabook.arabook.member.entity.Member;
+import com.arabook.arabook.review.entity.enums.ReviewTag;
 import com.arabook.arabook.storage.domain.common.entity.BaseTimeEntity;
-import com.arabook.arabook.storage.domain.review.entity.enums.ReviewTag;
 
 import lombok.AccessLevel;
 import lombok.Builder;
