+++ src/main/java/com/arabook/arabook/review/repository/ReviewCustomRepository.java
@@ -1,9 +1,9 @@
-package com.arabook.arabook.storage.domain.review.repository;
+package com.arabook.arabook.review.repository;
 
 import java.util.List;
 
-import com.arabook.arabook.api.review.controller.dto.response.ReviewDetailResponse;
-import com.arabook.arabook.api.review.controller.dto.response.ReviewResponse;
+import com.arabook.arabook.review.controller.dto.response.ReviewDetailResponse;
+import com.arabook.arabook.review.controller.dto.response.ReviewResponse;
 
 public interface ReviewCustomRepository {
   List<ReviewResponse> findReviewsByReviewerId(Long memberId);
